Moved InfluxDB and Kafka reporters to their own packages
