+++ gobblin-metrics/src/main/java/gobblin/metrics/kafka/ProducerCloseable.java
@@ -10,7 +10,7 @@
  * CONDITIONS OF ANY KIND, either express or implied.
  */
 
-package gobblin.metrics;
+package gobblin.metrics.kafka;
 
 import java.io.Closeable;
